Remove tcp stream spy from the tests
Little angel gone too soon, you were the shinest tool we had.
But the need for compatibility with Rust brings and end to this golden age of elegant testing.

R.I.P, you'll be missed :'(
