+++ tests/common/__init__.py
@@ -10,6 +10,5 @@ from .helpers import *  # noqa
 from .population import *  # noqa
 from .postgresql import *  # noqa
 from .smartcard import *  # noqa
-from .tcp import *  # noqa
 from .timeout import *  # noqa
 from .trio_clock import *  # noqa
